Move GHContext from gesture-handler to native-stack
