+++ src/native-stack/contexts/GHContext.tsx
@@ -1,5 +1,5 @@
 import React, { PropsWithChildren } from 'react';
-import { GestureProviderProps } from '../../native-stack/types';
+import { GestureProviderProps } from '../types';
 
 // context to be used when the user wants full screen swipe (see `gesture-handler` folder in repo)
 export const GHContext = React.createContext(
